refactor: filename authData (for consistency)
